requirements.txt
streamlit>=1.35.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
google-genai>=1.0.0
pdfplumber>=0.11.0
python-pptx>=1.0.0
